Added Gates to edit, update and destroy methods

+++ app/Http/Controllers/QuestionsController.php
@@ -67,6 +67,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            // abort(403, "You don't have access to that question!");
+            abort(403);
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -79,6 +83,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            // abort(403, "You're not allowed to update that question!");
+            abort(403, "Access denied");
+        }
         $question->update( $request->only(['title', 'body']) );
         return redirect(route('questions.index'))->with('success', 'Your question has been updated.');
     }
@@ -91,6 +99,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            // abort(403, "You're not allowed to delete that question!");
+            abort(403, "Access denied");
+        }
         $question->delete();
         return redirect(route('questions.index'))->with('success', "That question has now been deeted.");
     }
